+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.5.0
+  * Use `delegates` for message accessors/getters on `context`
+  * Added `context` delegates for `message`
+    * `fields` getter
+    * `properties` getter
+    * `content` accessor
+    * `messageAcked` accessor
+  * Added `context` delegates for `message.properties`
+    * `headers` getter
+  * Added `context` delegates for `message.fields`
+    * `exchange` accessor
+    * `routingKey` accessor
+    * `deliveryTag` getter
+    * `consumerTag` getter
+    * `redelivered` getter
+
 # 0.4.1
 Package:
   * Remove unused `ee-first`